+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.3] - 2021-01-16
+
 ### Added
 
 - FluidSynth is now reset when a "GM Mode On" or "GS Reset" SysEx message is received.
@@ -248,7 +250,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial version.
 
-[unreleased]: https://github.com/dwhinham/mt32-pi/compare/v0.8.2...HEAD
+[unreleased]: https://github.com/dwhinham/mt32-pi/compare/v0.8.3...HEAD
+[0.8.2]: https://github.com/dwhinham/mt32-pi/compare/v0.8.2..v0.8.3
 [0.8.2]: https://github.com/dwhinham/mt32-pi/compare/v0.8.1..v0.8.2
 [0.8.1]: https://github.com/dwhinham/mt32-pi/compare/v0.8.0..v0.8.1
 [0.8.0]: https://github.com/dwhinham/mt32-pi/compare/v0.7.1..v0.8.0